+++ application/views/layouts/account_layout.php
@@ -57,6 +57,9 @@
 <script src="<?= asset_url('assets/js/utils/string.js') ?>"></script>
 <script src="<?= asset_url('assets/js/utils/url.js') ?>"></script>
 <script src="<?= asset_url('assets/js/utils/validation.js') ?>"></script>
+<?php if (setting('altcha_enabled') === '1'): ?>
+<script src="<?= asset_url('assets/js/utils/altcha.js') ?>"></script>
+<?php endif; ?>
 <script src="<?= asset_url('assets/js/layouts/account_layout.js') ?>"></script>
 <script src="<?= asset_url('assets/js/http/account_http_client.js') ?>"></script>
 <script src="<?= asset_url('assets/js/http/localization_http_client.js') ?>"></script>