+++ assets/js/pages/login.js
@@ -21,6 +21,8 @@ App.Pages.Login = (function () {
     const $captchaText = $('.captcha-text');
     const $captchaTitle = $('.captcha-title');
     const $captchaHint = $('#captcha-hint');
+    const $altchaPayload = $('#altcha-payload');
+    const $altchaHint = $('#altcha-hint');
 
     /**
      * Refresh the captcha image.
@@ -53,14 +55,25 @@ App.Pages.Login = (function () {
                 return;
             }
         }
+        
+        if ($altchaPayload.length > 0 && $altchaPayload.val() === '') {
+            $altchaHint.text(lang('altcha_verification_failed')).fadeTo(400, 1);
+            
+            setTimeout(() => {
+                $altchaHint.fadeTo(400, 0);
+            }, 3000);
+            
+            return;
+        }
 
         const captcha = $captchaText.length > 0 ? $captchaText.val() : null;
+        const altchaPayloadValue = $altchaPayload.length > 0 ? $altchaPayload.val() : null;
 
         const $alert = $('.alert');
 
         $alert.addClass('d-none');
 
-        App.Http.Login.validate(username, password, captcha).done((response) => {
+        App.Http.Login.validate(username, password, captcha, altchaPayloadValue).done((response) => {
             if (response.captcha_verification === false) {
                 $captchaHint.text(lang('captcha_is_wrong')).fadeTo(400, 1);
 
@@ -74,6 +87,21 @@ App.Pages.Login = (function () {
 
                 return;
             }
+            
+            if (response.altcha_verification === false) {
+                $altchaHint.text(lang('altcha_verification_failed')).fadeTo(400, 1);
+                
+                setTimeout(() => {
+                    $altchaHint.fadeTo(400, 0);
+                }, 3000);
+                
+                // Reset ALTCHA widget
+                if (App.Utils.Altcha) {
+                    App.Utils.Altcha.reset('altcha-widget');
+                }
+                
+                return;
+            }
 
             if (response.success) {
                 window.location.href = vars('dest_url');
@@ -84,10 +112,22 @@ App.Pages.Login = (function () {
             }
         });
     }
+    
+    /**
+     * Initialize ALTCHA widget if present.
+     */
+    function initializeAltcha() {
+        if ($('#altcha-widget').length && App.Utils.Altcha) {
+            App.Utils.Altcha.initialize('altcha-widget');
+        }
+    }
 
     $loginForm.on('submit', onLoginFormSubmit);
 
     $captchaTitle.on('click', 'button', refreshCaptcha);
+    
+    // Initialize ALTCHA
+    initializeAltcha();
 
     return {};
 })();