+++ CHANGELOG.md
@@ -30,7 +30,7 @@ developers to maintain and readjust their custom modifications on the main proje
 - Allow customers to save their information in local storage (#1696)
 - Add a legal notice and imprint links to the booking page and the legal settings page (#1407)
 - Implementation of additional GDPR features in Easy!Appointments (#535)
-
+- ALTCHA integration added as an alternative CAPTCHA step (#1155)
 
 ### Changed